services/document-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==1.4.41
asyncpg==0.27.0
networkx==3.2.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3